+++ transpiler/examples/contacts.expected.dart
@@ -163,7 +163,7 @@ class _ContactDetailScreenState extends State<ContactDetailScreen> {
   void toggle_fav() {
     dynamic c = shared.contacts.cast<dynamic>().firstWhere((c) => c['name'] == widget.contact['name'], orElse: () => null);
     shared.update(() {
-      shared.contacts = shared.contacts.map((e) => e == c ? {...c, 'favourite': !c['favourite']} : e).toList();
+      shared.contacts = shared.contacts.map((e) => e == c ? <String, dynamic>{...(c as Map), 'favourite': !c['favourite']} : e).toList();
     });
   }