+++ transpiler/examples/type-hints.expected.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
 class SharedState extends ChangeNotifier {
-  List<Item> items = <dynamic>[{'name': 'Milk'}, {'name': 'Bread'}];
+  List<dynamic> items = <dynamic>[{'name': 'Milk'}, {'name': 'Bread'}];
 
   void update(void Function() fn) {
     fn();